Replaced _ in the item name with a space
